# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_adapter_presenter.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Amplified Access
  description: Amplified Access Watchtower — incident reporting and monitoring platform
    API.
  title: Watchtower API
  version: "1.0"
paths:
...
  /alerts:
    get:
      description: Returns all alert subscriptions for the given email address
      parameters:
      - description: Subscriber email address
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: Get subscriptions by email
      tags:
      - Alerts
    post:
      consumes:
      - application/json
      description: Subscribes an email address to incident alerts for specified types
        and locations
      parameters:
      - description: Subscription payload
        in: body
        name: body
        required: true
        schema:
          properties:
            alertFrequency:
              type: string
            email:
              type: string
            emailNotifications:
              type: boolean
            incidentTypes:
              items:
                type: string
              type: array
            locations:
              items:
                type: object
              type: array
            name:
              type: string
            phone:
              type: string
            preferredLanguage:
              type: string
            severityLevels:
              items:
                type: string
              type: array
            smsNotifications:
              type: boolean
            timezone:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: Create alert subscription
      tags:
      - Alerts
  /alerts/{id}:
    patch:
      consumes:
      - application/json
      description: Updates preferences for an existing alert subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          properties:
            alertFrequency:
              type: string
            emailNotifications:
              type: boolean
            incidentTypes:
              items:
                type: string
              type: array
            locations:
              items:
                type: object
              type: array
            name:
              type: string
            phone:
              type: string
            preferredLanguage:
              type: string
            severityLevels:
              items:
                type: string
              type: array
            smsNotifications:
              type: boolean
            timezone:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: Update alert subscription
      tags:
      - Alerts
  /alerts/{id}/activate:
    post:
      description: Resumes alert delivery for a previously deactivated subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: Activate alert subscription
      tags:
      - Alerts
  /alerts/{id}/deactivate:
    post:
      description: Pauses alert delivery for the given subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: Deactivate alert subscription
      tags:
      - Alerts
...
  /superadmin/alerts/active:
    get:
      description: Returns every active subscription across the platform (super admin
        only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: List all active alert subscriptions
      tags:
      - Super Admin – Alerts
  /superadmin/alerts/stats:
    get:
      description: Returns aggregate statistics for all alert subscriptions (super
        admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: Get alert subscription statistics
      tags:
      - Super Admin – Alerts
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token obtained from authentication. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"